+++ marker/views/company.py
@@ -1,3 +1,4 @@
+import datetime
 import logging
 
 import pycountry
@@ -203,6 +204,8 @@ class CompanyView:
         NIP = self.request.params.get("NIP", None)
         REGON = self.request.params.get("REGON", None)
         KRS = self.request.params.get("KRS", None)
+        date_from = self.request.params.get("date_from", None)
+        date_to = self.request.params.get("date_to", None)
         _sort = self.request.params.get("sort", "created_at")
         _order = self.request.params.get("order", "desc")
         sort_criteria = dict(SORT_CRITERIA_COMPANIES)
@@ -265,6 +268,16 @@ class CompanyView:
         if KRS:
             stmt = stmt.filter(Company.KRS == KRS)
             q["KRS"] = KRS
+
+        if date_from:
+            date_from_dt = datetime.datetime.strptime(date_from, "%Y-%m-%dT%H:%M")
+            stmt = stmt.filter(Company.created_at >= date_from_dt)
+            q["date_from"] = date_from
+
+        if date_to:
+            date_to_dt = datetime.datetime.strptime(date_to, "%Y-%m-%dT%H:%M")
+            stmt = stmt.filter(Company.created_at <= date_to_dt)
+            q["date_to"] = date_to
 
         if view_mode == "contacts":
             q["view"] = "contacts"
@@ -1072,6 +1085,8 @@ class CompanyView:
         subdivision = [
             value for value in self.request.params.getall("subdivision") if value
         ]
+        date_from = self.request.params.get("date_from", None)
+        date_to = self.request.params.get("date_to", None)
         _sort = self.request.params.get("sort", "shared_tags")
         _order = self.request.params.get("order", "desc")
         colors = dict(COLORS)
@@ -1119,6 +1134,16 @@ class CompanyView:
         if subdivision:
             stmt = stmt.filter(Company.subdivision.in_(subdivision))
             q["subdivision"] = list(subdivision)
+
+        if date_from:
+            date_from_dt = datetime.datetime.strptime(date_from, "%Y-%m-%dT%H:%M")
+            stmt = stmt.filter(Company.created_at >= date_from_dt)
+            q["date_from"] = date_from
+
+        if date_to:
+            date_to_dt = datetime.datetime.strptime(date_to, "%Y-%m-%dT%H:%M")
+            stmt = stmt.filter(Company.created_at <= date_to_dt)
+            q["date_to"] = date_to
 
         q["sort"] = _sort
         q["order"] = _order